+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Autoscaler Change Log
 
 - Fix filtering pods and add custom setting HF_VAR_autoscalerJobLabel
 - Fix missing intialization set in GCP provider
+- Fix getDemand - nodeName is no longer required
 
 0.4.0 October 22, 2020
 ------------------------