+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Autoscaler Change Log
 - Fix filtering pods and add custom setting HF_VAR_autoscalerJobLabel
 - Fix missing intialization set in GCP provider
 - Fix getDemand - nodeName is no longer required
+- Fix scalingOptimizer - use correct demand frames before/after scaling
 
 0.4.0 October 22, 2020
 ------------------------