+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Autoscaler Change Log
 0.5.0 under development
 ------------------------
 
+
+0.4.1 October 23, 2020
+------------------------
+
 - Fix filtering pods and add custom setting HF_VAR_autoscalerJobLabel
 - Fix missing intialization set in GCP provider
 - Fix getDemand - nodeName is no longer required